Issue #10 refactor `tools::fill` - Fixes #10
